fix: move buildSchemaDirectives to its own file
